requirements.txt
annotated-types==0.5.0
anyio==4.0.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
pydantic==2.3.0
pydantic-core==2.6.3
PyJWT==2.8.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
